% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compile}
\alias{compile}
\title{Extract bluefish MRIP data}
\usage{
compile(type, dir = "MRIP_data", outdir = "data-raw", species = "BLUEFISH")
}
\arguments{
\item{type}{The type of MRIP data to compile. One of c("effort", "catch")}

\item{dir}{The directory with the spreadsheets}

\item{outdir}{Where to save the .RDS file}

\item{species}{The species common name as it appears in the MRIP data (probably in all caps)}
}
\value{
Saves the intermediate .RDS object of compiled MRIP effort or catch
}
\description{
This function compiles MRIP data from individual spreadsheets into an intermediate .RDS file
}
