% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bluefish_bottom_trawl.R
\name{create_bluefish_bottom_trawl}
\alias{create_bluefish_bottom_trawl}
\title{Create bluefish NMFS bottom trawl indicators}
\usage{
create_bluefish_bottom_trawl(
  data = readRDS(here::here("data-raw/nmfs_survey_expanded_9c.RDS")),
  method = "gam",
  return = TRUE
)
}
\arguments{
\item{data}{The full bluefish survey data}

\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Saves the R data object `bluefish_bottom_trawl`
}
\description{
This function creates bluefish NMFS bottom trawl indicators: min, mean, and max latitude, longitude, and temperature
}
