% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rec_indicators.R
\name{create_bluefish_trips}
\alias{create_bluefish_trips}
\title{Create bluefish trips indicator}
\usage{
create_bluefish_trips(files, return = TRUE)
}
\arguments{
\item{files}{A list of the full file names of annual directed trip data (.csv format)}

\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Updates the package object `bluefish_trips`

Directed trip data must be downloaded manually from the MRIP query tool
https://www.fisheries.noaa.gov/data-tools/recreational-fisheries-statistics-queries
2022 bluefieh ESP: downloaded atlantic coast by state, all modes by mode,
all waves, all areas by area;
selected all bluefish associated trips - if it was a target species, if it was
caught, if it was released
}
\description{
This function creates the bluefish trips indicator
}
