% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_comm_indicators.R
\name{create_comm_indicators}
\alias{create_comm_indicators}
\title{Create commercial indicators}
\usage{
create_comm_indicators(
  data = here::here("data-raw/share/SOCIEOECONOMIC_COMMERCIAL_INDICATORS_FINAL.csv"),
  return = TRUE
)
}
\arguments{
\item{data}{Filepath to the indicator data}

\item{return}{Boolean. Whether to return the indicators as an object in the global environment}
}
\value{
Saves R object `comm_indicators`, returns commercial indicators
}
\description{
This function transforms bluefish commercial indicators from an input spreadsheet to an R object
}
