% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_community_indicators.R
\name{create_community_indicators}
\alias{create_community_indicators}
\title{Create community indicators}
\usage{
create_community_indicators(
  filepath = here::here("data-raw/2009-2019 Rec Scores_ME_MS_031522 cw.xlsx"),
  return = TRUE
)
}
\arguments{
\item{return}{Boolean. Whether to return the indicator as an object in the global environment}

\item{data}{The engagement and reliance data}
}
\value{
Saves the R data object `community_indicators`
}
\description{
This function creates community indicators: mean regional engagement and mean regional reliance
}
