% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_temp_indicators.R
\name{create_july_proportion}
\alias{create_july_proportion}
\title{Create July temperature indicator}
\usage{
create_july_proportion(data, return = TRUE)
}
\arguments{
\item{data}{Filepath to the indicator data}

\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Saves R object `july_proportion`, returns temperature indicator
}
\description{
This function transforms a bluefish temperature indicator from an input spreadsheet to an R object
}
