% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mako_iccat.R
\name{create_mako}
\alias{create_mako}
\title{Create mako shark indicator}
\usage{
create_mako(file = here::here("data-raw/SMA2017_Bratio.xlsx"), return = TRUE)
}
\arguments{
\item{file}{A .xlsx with the ICCAT mako population model outputs}

\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Saves the R data object `mako`
}
\description{
This function creates the mako shark indicator
}
