% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mrip_estimated.R
\name{create_mrip_catch}
\alias{create_mrip_catch}
\title{Create MRIP *estimated* catch}
\usage{
create_mrip_catch(rds = here::here("data-raw/catch_wave.RDS"), return = TRUE)
}
\arguments{
\item{rds}{The .RDS file with the MRIP estimated catch data}
}
\value{
Updates the package object `mrip_catch`
}
\description{
This function creates MRIP estimated catch from a .RDS file
}
