% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mrip_raw.R
\name{create_mrip_catch_raw}
\alias{create_mrip_catch_raw}
\title{Create MRIP *raw* catch}
\usage{
create_mrip_catch_raw(rds = here::here("data-raw/catch.RDS"))
}
\arguments{
\item{rds}{The .RDS file with the MRIP raw catch data}
}
\value{
Updates the package object `mrip_catch_raw`
}
\description{
This function creates MRIP raw catch from a .RDS file
}
