% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mrip_estimated.R
\name{create_mrip_effort}
\alias{create_mrip_effort}
\title{Create MRIP *estimated* effort}
\usage{
create_mrip_effort(rds = here::here("data-raw/effort.RDS"), return = TRUE)
}
\arguments{
\item{rds}{The .RDS file with the MRIP estimated effort data}
}
\value{
Updates the package object `mrip_effort`
}
\description{
This function creates MRIP estimated effort from a .RDS file
}
