% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mrip_raw.R
\name{create_mrip_trip_raw}
\alias{create_mrip_trip_raw}
\title{Create MRIP *raw* trips}
\usage{
create_mrip_trip_raw(rds = here::here("data-raw/trip.RDS"))
}
\arguments{
\item{rds}{The .RDS file with the MRIP raw trip data}
}
\value{
Updates the package object `mrip_trip_raw`
}
\description{
This function creates MRIP raw trips from a .RDS file
}
