% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_temp_indicators.R
\name{create_n_18_day}
\alias{create_n_18_day}
\title{Create number of 18 degree days temperature indicator}
\usage{
create_n_18_day(data, return = TRUE)
}
\arguments{
\item{data}{Filepath to the indicator data}

\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Saves R object `n_18_day`, returns temperature indicator
}
\description{
This function transforms a bluefish temperature indicator from an input spreadsheet to an R object
}
