% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rec_indicators.R
\name{create_prop_bluefish_trips}
\alias{create_prop_bluefish_trips}
\title{Create proportion trips targeting bluefish indicator}
\usage{
create_prop_bluefish_trips(
  total = mrip_effort,
  bluefish = bluefish_trips,
  return = TRUE
)
}
\arguments{
\item{total}{The mrip trip data (R object `mrip_effort`)}

\item{bluefish}{The bluefish directed trips (R object `bluefish_trips`)}

\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Saves the R data object `prop_bluefish_trips`
}
\description{
This function creates the proportion trips targeting bluefish indicator
}
