% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_relative_condition.R
\name{create_relative_condition}
\alias{create_relative_condition}
\title{Create relative condition}
\usage{
create_relative_condition(return = TRUE)
}
\arguments{
\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Saves the R data object `relative_condition`
}
\description{
This function creates relative condition indicators
}
