% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rec_indicators.R
\name{create_total_rec_catch}
\alias{create_total_rec_catch}
\title{Create total recreational catch indicator}
\usage{
create_total_rec_catch(data = mrip_catch, return = TRUE)
}
\arguments{
\item{data}{The mrip data (R object `mrip_catch`), already subset to bluefish only}

\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Saves the R data object `mrip_catch`
}
\description{
This function creates the total recreational catch indicator
}
