% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rec_indicators.R
\name{create_total_rec_landings}
\alias{create_total_rec_landings}
\title{Create bluefish recreational landings indicator}
\usage{
create_total_rec_landings(data = mrip_catch, return = TRUE)
}
\arguments{
\item{data}{The mrip catch data (R object `mrip_catch`)}

\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Saves the R data object `total_rec_landings`
}
\description{
This function creates the bluefish recreational landings indicator
}
