% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wind_indicators.R
\name{create_wind_indicators}
\alias{create_wind_indicators}
\title{Create wind indicators}
\usage{
create_wind_indicators(
  vwind_file,
  uwind_file,
  months = c(paste0("X", 1979:2021, ".04.01"), paste0("X", 1979:2021, ".05.01")),
  return = TRUE
)
}
\arguments{
\item{vwind_file}{Filepath to the wind data}

\item{uwind_file}{Filepath to the wind data}

\item{months}{Months to use}

\item{return}{Boolean. Whether to return the indicators as an object in the global environment}
}
\value{
Saves R object `wind_indicators`, returns wind indicators
}
\description{
This function calculates bluefish wind indicators from input .nc files
Monthly 10m wind downloaded from PSL: https://downloads.psl.noaa.gov/Datasets/NARR/Monthlies/monolevel/
}
