% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_and_extract}
\alias{download_and_extract}
\title{Download MRIP data}
\usage{
download_and_extract(url, pattern = "ps_.{1,}", exdir = "MRIP_data")
}
\arguments{
\item{url}{A list of the URLs to download}
}
\value{
Downloads MRIP data
}
\description{
This function downloads MRIP data from a URL and unzips it into the specified folder
}
