% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp_traffic_fig.R
\name{esp_traffic_fig}
\alias{esp_traffic_fig}
\title{Plot indicator traffic light figure}
\usage{
esp_traffic_fig(
  data,
  name,
  out = "one_pager",
  paginate = FALSE,
  label = TRUE,
  status = TRUE,
  caption = "",
  ncolumn = 1,
  silent = FALSE,
  min_year = 1985,
  ...
)
}
\arguments{
\item{data}{The ESP indicator data. Should have a column for Year and a column with indicator values.}

\item{name}{The file name for the image. Will be saved relative to the working directory.}

\item{out}{Whether the function should save the plot, print a ggplot object in markdown, or return a ggplot object to the working environment (for use with `one_pager()`). One of c("ggplot", "save", "one_pager")}

\item{paginate}{Whether to paginate the plots with `ggforce::facet_wrap_paginate`}

\item{label}{Whether to label the facets with a, b, c, etc.}

\item{status}{Whether to label the facets with the indicator status}

\item{caption}{A caption for the figure}

\item{ncolumn}{How many columns the figure should have (1 by default)}

\item{silent}{Whether to print the caption}

\item{min_year}{The minimum year to show on the plots. If left NULL (the default), the minimum year will be the first year of the dataset.}

\item{...}{Passed to `ggplot2::ggsave`}
}
\value{
An image file
}
\description{
This function plots ESP indicator traffic light figures
}
