% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_combine}
\alias{test_combine}
\title{Combine indicator data}
\usage{
test_combine(...)
}
\arguments{
\item{...}{R objects}
}
\value{
Combined indicator data
}
\description{
This function combines indicator data
}
